added adaptive and tidied up a bit
