+++ src/app/layout.tsx
@@ -1,9 +1,11 @@
 import './globals.css';
 import type { Metadata } from 'next';
 import { Manrope } from 'next/font/google';
+import { Open_Sans } from 'next/font/google';
 import Footer from '@/components/footer';
 
 const manrope = Manrope({ subsets: ['latin', 'cyrillic'] });
+const openSans = Open_Sans({ subsets: ['latin', 'cyrillic'] });
 
 export const metadata: Metadata = {
   title: 'Create Next App',
@@ -13,7 +15,8 @@ export const metadata: Metadata = {
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en">
-      <body className={`${manrope.className} bg-white`}>{children}
+      <body className={`${manrope.className} ${openSans.className} bg-white`}>
+        {children}
         <section className="mb-10 mt-[90px]">
           <Footer />
         </section>
